requirements.txt
django==5.0.2
djangorestframework==3.15.2
ipython==8.32.0
environs
httpie
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
django-crispy-forms
crispy-bootstrap5
djangorestframework-oauth
django-oauth-plus oauth2
django-oauth2-provider
